requirements.txt
pyqubo==1.5.0
dimod==0.12.18
dwave-neal==0.6.0
# openjij is optional
# openjij==0.6.7
